feat: Add variant unwrapping methods
This commit adds methods for unwrapping enum variants,
inline with the implementations of the `Result` and `Option` types,
providing methods to concisely and consistently implement handling
for enum variants in cases where doing so is trivial.

This includes basic methods which will panic on invalid variants,
methods which return an alternative value, and methods
with alternative logic.